+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.3.0
 
 - **BREAKING**: Fix type errors for a list of custom items.
-- **BREAKING**: `fromJson` and `toJson` require `Object` instead of `Map<String, Object?>`.
+- **BREAKING**: `fromJson` and `toJson` require `dynamic` instead of `Map<String, Object?>`.
 
 ## 0.2.0